[feat] 新增MACD和布林带衍生特征，修复模型推理评测时的精度缺陷，优化快速跳过已最新的股票数据的逻辑
1. 新增数据库表字段与迁移逻辑，支持MACD/布林带特征存储
2. 实现EMA、MACD、布林带特征计算逻辑
3. 优化快速跳过已最新的股票数据的逻辑
4. 修复模型推理评测时的精度缺陷

+++ src/training_utils.py
@@ -401,7 +401,8 @@ def evaluate_model(model, eval_inputs, eval_targets, eval_cumulative_returns,
             batch_inputs = torch.tensor(eval_inputs[start_idx:end_idx], dtype=torch.float32, device=device)
             with amp_ctx:
                 logits = model(batch_inputs)
-            scores = torch.sigmoid(logits.float())
+            logits = logits.float()
+            scores = torch.sigmoid(logits)
             all_preds.append(scores.cpu().numpy().flatten())
 
             if criterion is not None:
